+++ changelog.md
@@ -11,6 +11,7 @@ All notable changes to Lock will be documented in this file. This file follows t
 ### Fixed
 
 - Fixed a bug where only array data could be passed to the `PermissionFactory`
+- Fixed a bug where a action alias got denied but was still accepted
 
 ## 1.0.0-alpha.2 - 2014-12-05